import restructuring, begin of test generalization

+++ src/spflow/torch/sampling/general/layers/leaves/parametric/cond_log_normal.py
@@ -95,6 +95,6 @@ def sample(
 
         data[torch.meshgrid(sampling_ids, torch.tensor(node_scope.query), indexing="ij")] = (
             layer.dist(mean=mean, std=std, node_ids=[node_id]).sample((sampling_mask.sum(),)).to(mean.device)
-        )
+        ).type(torch.float64)
 
     return data
